requirements.txt
flask==3.0.2
werkzeug==3.0.1
python-docx==1.1.0
openai==1.12.0
python-dotenv==1.0.1
boto3==1.38.11 